20 Hour
made it so you can confirm a craft and it will also play animation. It doesnt actually give anything yet. You can also now exit the crafting ui. I added the rest of the units to the json and made it so it defaults to pepper if the unit doesnt exist when it tries to summon. Made function to actually get crafted units.