36 Hour 31 minute
I forgot to commit at 36 hours and only noticed 20 minutes into the hour, so i decided to just commit at the 30. I added a new click sound effect for burnt toast that sounds a lot nicer to the ears (hopefully), and did some effects in godot to it. I also fixed a major bug that would crash the game if you end up with 0 engrams while crafting.