54 hour
I added a new sprite for blender, oven, and rolling pin. I decided to rescale the size of the game. This is so I can just have higher quality textures and not a half-baked pixel effect. Also gives more room for text which is very necessary for a game like this. Started moving around ui to fix.